requirements.txt
dgl-cu116==0.9.1
numpy==1.23.4
ogb==1.3.4
torch==1.12.1+cu116
scikit-learn==1.1.3
tqdm==4.64.1

